requirements.txt
aiodns==3.0.0
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
frozenlist==1.3.3
idna==3.4
meross-iot==0.4.5.9
multidict==6.0.4
paho-mqtt==1.6.1
pycares==4.3.0
pycparser==2.21
python-dotenv==0.21.1
requests==2.28.2
urllib3==1.26.14
yarl==1.8.2
